👌 IMPROVE: refactor
